requirements.txt
numpy
torch
torchvision
starlette
uvicorn
python-multipart
aiofiles
python-dotenv
aiohttp
Pillow
